pyproject.toml
[tool.poetry]
name = "libbiblio"
version = "2.0.0a"
description = ""
authors = [
    "Yuri Chiucconi <yuri.chiucconi@exact-lab.it>",
    "Jacopo Nespolo <jacopo.nespolo@exact-lab.it>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.8.2"
# xsdata = {extras = ["cli"], version = "^21.11"}
xsdata = "^21.11"
asyncclick = "^8.0.1.3"
anyio = "^3.4.0"
poetry-dotenv-plugin = "^0.1.0-alpha.2"
Sickle = "^0.7.0"
alembic = "^1.7.5"
mypy = "^0.910"
SQLAlchemy = {extras = ["mypy"], version = "^1.4.27"}
psycopg2-binary = "^2.9.2"
sqlalchemy-stubs = "^0.4"
lxml = "^4.7.1"
mkdocs = "^1.3.1"
mkdocstrings = {extras = ["python"], version = "^0.19.0"}
mkdocs-material = "^8.3.9"
mkdocs-toc-sidebar-plugin = "^0.1.0"
mdx-include = "^1.4.1"
mkdocs-markdownextradata-plugin = "^0.2.5"
appdirs = "^1.4.4"
sqlmodel = "^0.0.8"
xmltodict = "^0.13.0"
typer = {extras = ["all"], version = "^0.7.0"}

# [tool.poetry.scripts]
# biblio = 'libbiblio.main:biblio'

[tool.poetry.group.dev.dependencies]
memray = "^1.7.0"
ipython = "^8.3.0"
pytest = "^6.2.5"
pytest-asyncio = "^0.16.0"
devtools = "^0.8.0"
ipykernel = "^6.15.1"
pandas = "^1.4.3"
matplotlib = "^3.5.2"
seaborn = "^0.11.2"
python-dotenv = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
